refector : 로또번호 맞추기 로직 변경

+++ src/main/java/autolotto/model/ProfitMatcher.java
@@ -14,7 +14,7 @@ public class ProfitMatcher {
         }
     };
 
-    public static float getMatchProfit(int matchingNumber, int matchingCount){
+    public static int getMatchProfit(int matchingNumber, int matchingCount){
         return matchProfitList.get(matchingNumber) * matchingCount;
     }
